minor improvement

+++ modules/GPS_module/src/GPS_module_1.h
@@ -3,7 +3,7 @@
 
 #include "GPS_module_types.h"
 
-extern int GPS_driver_obtain_current_position(uint8_t * position_as_bytes, uint8_t * hmac_as_bytes);
+extern uint8_t GPS_driver_obtain_current_position(uint8_t * position_as_bytes, uint8_t * hmac_as_bytes);
 
 enum GPS_return_status init_crypto_module();
